Move PDB reader functions into mpnn_embeddings module
Consolidate the PDB/CIF parsing functions directly into mpnn_embeddings.py
instead of a separate pdb_reader module:

- Move _get_inputs_mpnn, _get_structure, _compute_cb, _np_extend, _np_norm
  from pdb_reader.py into mpnn_embeddings.py as private functions
- Delete pdb_reader.py module
- Update tests to mock _get_inputs_mpnn directly
- Add comparison test verifying our implementation matches softalign output
